pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cmc-framework"
version = "0.2.0"
dependencies = [
    "psycopg2-binary",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]

[project.scripts]
cmc = "core.cli:main"